Move read_with_python_encoding
