+++ mypy/util.py
@@ -2,6 +2,7 @@
 
 import re
 import subprocess
+import hashlib
 from xml.sax.saxutils import escape
 from typing import TypeVar, List, Tuple, Optional, Sequence, Dict
 
@@ -58,6 +59,49 @@ def find_python_encoding(text: bytes, pyversion: Tuple[int, int]) -> Tuple[str, 
     else:
         default_encoding = 'utf8' if pyversion[0] >= 3 else 'ascii'
         return default_encoding, -1
+
+
+class DecodeError(Exception):
+    """Exception raised when a file cannot be decoded due to an unknown encoding type.
+
+    Essentially a wrapper for the LookupError raised by `bytearray.decode`
+    """
+
+
+def read_with_python_encoding(path: str, pyversion: Tuple[int, int]) -> Tuple[str, str]:
+    """Read the Python file with while obeying PEP-263 encoding detection.
+
+    Returns:
+      A tuple: the source as a string, and the hash calculated from the binary representation.
+    """
+    source_bytearray = bytearray()
+    encoding = 'utf8' if pyversion[0] >= 3 else 'ascii'
+
+    with open(path, 'rb') as f:
+        # read first two lines and check if PEP-263 coding is present
+        source_bytearray.extend(f.readline())
+        source_bytearray.extend(f.readline())
+        m = hashlib.md5(source_bytearray)
+
+        # check for BOM UTF-8 encoding and strip it out if present
+        if source_bytearray.startswith(b'\xef\xbb\xbf'):
+            encoding = 'utf8'
+            source_bytearray = source_bytearray[3:]
+        else:
+            _encoding, _ = find_python_encoding(source_bytearray, pyversion)
+            # check that the coding isn't mypy. We skip it since
+            # registering may not have happened yet
+            if _encoding != 'mypy':
+                encoding = _encoding
+
+        remainder = f.read()
+        m.update(remainder)
+        source_bytearray.extend(remainder)
+        try:
+            source_text = source_bytearray.decode(encoding)
+        except LookupError as lookuperr:
+            raise DecodeError(str(lookuperr))
+        return source_text, m.hexdigest()
 
 
 _python2_interpreter = None  # type: Optional[str]
